feat: add search function in my-groups page
