--- apps/dashboard/src/types/groups/group.ts
@@ -1,6 +1,0 @@
-export type Group = {
-    name: string
-    description: string
-    treeDepth: number
-    members: string[]
-}
